feat(money): migrate all money to integer paise with a typed Money utility (GAP-9/US-401)
Money was whole-rupee `number` math over DECIMAL(10,2) columns — fine today, but
imprecise as finance grows and blocking for GST/paise. This migrates every money
field and calculation to integer paise end-to-end, behind a typed Money utility,
with zero change to any visible rupee value.

Storage: all 19 money columns Decimal(10,2) → BIGINT paise, converted in place via
`USING round(col*100)::bigint` (a bare cast would truncate). The exclusion
constraint is untouched (the spurious `stay DROP DEFAULT` was stripped, constraint
verified intact after apply). Percentages (commission/adjust pct) stay Decimal.

Representation: `src/lib/money.ts` — Money is a branded integer-paise number, exact
below 2^53. Rupees↔paise conversions happen ONLY at input parsing and display;
everything between (DB, wire, calc) is paise. Prisma returns bigint for money, so a
BigInt#toJSON shim serialises it as a number for API responses. formatPaise renders
whole rupees, byte-identical to the old displayINR; commission/refund keep
whole-rupee rounding (pctOfMoneyWholeRupee) so no finance value shifts.

Surface (49 files): finance/pricing/cancellation/agents/tours calc; reservations/
payments/refunds/payouts/expenses/room-type/pricing-override routes (Zod → int
paise); every money form (rupees→paise on submit, paise→rupees on prefill);
displays via the now paise-aware displayINR/displayMoney; UPI QR/link amounts
(rupees); CSV exports (rupees, byte-identical); seed + demo-data (paise).

Tests: property-based money.test.ts (fast-check, dev-only) for the arithmetic
invariants; a CSV parity test proving the accountant CSVs still emit whole rupees
(2500, not 250000); existing finance/pricing/cancellation/agents/payout tests
updated to paise semantics.

+++ src/app/api/reservations/[id]/route.ts
@@ -33,8 +33,8 @@ const updateSchema = z
     status: z.enum(["confirmed", "cancelled", "no_show"]).optional(),
     arrivalTime: z.string().nullable().optional(),
     specialRequests: z.string().nullable().optional(),
-    grossAmount: z.number().nonnegative().nullable().optional(),
-    advanceRequired: z.number().nonnegative().nullable().optional(),
+    grossAmount: z.number().int().nonnegative().nullable().optional(), // paise (GAP-9)
+    advanceRequired: z.number().int().nonnegative().nullable().optional(), // paise
     otaRef: z.string().nullable().optional(),
     // Optimistic-concurrency token the edit form round-trips (see L-4).
     expectedVersion: z.number().int().optional(),
